requirements.txt
Django==5.2.9
djangorestframework==3.16.1
dj-rest-auth==7.0.1
django-allauth==65.13.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
django-environ==0.12.0
psycopg2-binary==2.9.11
Pillow==12.0.0

black==25.12.0
